requirements.txt
djangorestframework
django-rest-swagger
drf-yasg